:lipstick: Add 홈화면 접속 시에 로딩 컴포넌트 렌더링
